+++ src/Control/Effect/State/Strict.hs
@@ -7,6 +7,9 @@ module Control.Effect.State.Strict
 , evalState
 , execState
 , StateC(..)
+  -- * Re-exports
+, Carrier
+, run
 ) where
 
 import Control.Applicative (Alternative(..))